extract records
