+++ rust/src/benchmark/non_cs_counter.rs
@@ -11,7 +11,8 @@ use std::{
     time::Duration,
 };
 
-use crate::benchmark::{helper::create_zstd_writer, Record};
+use crate::benchmark::records::Record;
+use crate::benchmark::{helper::create_zstd_writer, records};
 
 use histo::Histogram;
 use libdlock::{
