+++ rust/src/benchmark/one_three_ratio_counter.rs
@@ -10,7 +10,8 @@ use std::{
     time::Duration,
 };
 
-use crate::benchmark::{helper::create_zstd_writer, Record};
+use crate::benchmark::records::Record;
+use crate::benchmark::{helper::create_zstd_writer, records};
 
 use histo::Histogram;
 use libdlock::{
