+++ rust/src/benchmark/subversion_job.rs
@@ -21,8 +21,9 @@ use libdlock::{
 use thread_priority::{ThreadPriority, ThreadPriorityValue};
 
 use crate::benchmark::helper::create_zstd_writer;
+use crate::benchmark::records::Record;
 
-use super::{bencher::LockBenchInfo, Record};
+use super::{bencher::LockBenchInfo, records};
 
 thread_local! {
     static WRITER: OnceCell<RefCell<Writer<AutoFinishEncoder<'static, File, Box<dyn FnMut(Result<File, std::io::Error>) + Send>>>>> = OnceCell::new();
